use Int for monthly turnover

+++ src/main/kotlin/gym/reporting/use_cases/TurnoverForAGivenMonth.kt
@@ -5,10 +5,10 @@ import gym.subscriptions.domain.SubscriptionEventStore
 class TurnoverForAGivenMonth(
     private val subscriptionEventStore: SubscriptionEventStore
 ) {
-    fun handle(command: TurnoverForAGivenMonthQuery): Double {
+    fun handle(command: TurnoverForAGivenMonthQuery): Int {
 
         return subscriptionEventStore.onGoingSubscriptions(command.asOfDate)
             .map { it.monthlyTurnover() }
-            .fold(0.0, { sum, monthlyTurnover -> sum + monthlyTurnover })
+            .fold(0, { sum, monthlyTurnover -> sum + monthlyTurnover })
     }
 }
